Refactored star bar and added semantics #25
