+++ lib/game_board/presentation/games_browse/widgets/body/games_browse_sliver_grid.dart
@@ -1,9 +1,9 @@
 import 'package:extended_image/extended_image.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:zacatrusa/core/optional.dart';
 import 'package:zacatrusa/game_board/presentation/core/routing/games_router_delegate.dart';
+import 'package:zacatrusa/game_board/presentation/core/widgets/star_bars_indicator.dart';
 
 import '../../../../../constants/app_margins.dart';
 import '../../../../zacatrus/domain/browse_page/game_overview.dart';
@@ -67,7 +67,11 @@ class ListGameItem extends ConsumerWidget {
                 children: [
                   GameName(game: game),
                   if (game.numberOfComments != null) Comments(game: game),
-                  if (game.stars != null) Stars(game: game),
+                  if (game.stars != null)
+                    Padding(
+                        padding:
+                            const EdgeInsets.only(top: innerElementsPadding),
+                        child: StarsBarIndicator(stars: game.stars!)),
                   if (game.price != null) Price(game: game),
                 ],
               ))
@@ -130,30 +134,6 @@ class Comments extends StatelessWidget {
         game.numberOfComments!.toString() +
             (game.numberOfComments! > 1 ? " comentarios" : " comentario"),
         style: Theme.of(context).textTheme.caption,
-      ),
-    );
-  }
-}
-
-class Stars extends StatelessWidget {
-  const Stars({
-    Key? key,
-    required this.game,
-  }) : super(key: key);
-
-  final GameOverview game;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(top: innerElementsPadding),
-      child: RatingBarIndicator(
-        itemBuilder: (context, index) => const Icon(
-          Icons.star,
-          color: Colors.amber,
-        ),
-        rating: game.stars!,
-        itemSize: 25.0,
       ),
     );
   }
